change condition to enter if block
Change if condition such that block is entered
only when the err is not nil and its not due to no result
being found. In some handlers this appeared as if there was an omissiong
of error handling, when it's ok to follow through, when no results
are tolerated (no feeds yet).